agents/requirements.txt
# Core (required)
web3>=6.0.0
python-dotenv>=1.0.0
websockets>=12.0
requests>=2.28.0

# API Server
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0

# LLM Providers (install at least one)
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0

